Make separation between library and examples
Provide the `examples/` directory with examples and add `--example`
argument to `cargo xtask run`.
