+++ examples/cli/examples/cli/file_open.rs
