+++ examples/cli/examples/cli/socket_connect.rs
