+++ examples/daemon/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "daemon"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow = { version = "1", features = ["backtrace"] }
+clap = { version = "4.2", features = ["derive"] }
+env_logger = "0.10"
+guardity = { path = "../../guardity" }
+log = "0.4"
+tokio = { version = "1.25", features = ["macros", "rt", "rt-multi-thread", "net", "signal", "sync"] }
