+++ examples/daemon/examples/daemon.rs
