--- guardctl/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "guardctl"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-anyhow = { version = "1.0", features = ["backtrace"] }
-aya = "0.11"
-clap = { version = "4.1", features = ["derive"] }
-cli-table = "0.4"
-guardity = { path = "../guardity" }
-guardity-common = { path = "../guardity-common", features = ["user"] }
-tokio = { version = "1.25", features = ["macros", "rt", "rt-multi-thread", "net", "signal", "sync"] }
