+++ guardity/Cargo.toml
@@ -2,14 +2,13 @@
 name = "guardity"
 version = "0.1.0"
 edition = "2021"
-publish = false
 
 [dependencies]
+anyhow = { version = "1", features = ["backtrace"] }
 aya = { git = "https://github.com/aya-rs/aya", branch = "main", features=["async_tokio"] }
 bytes = "1.4"
 clap = { version = "4.2", features = ["derive"] }
 guardity-common = { path = "../guardity-common", features = ["user"] }
-anyhow = { version = "1", features = ["backtrace"] }
 env_logger = "0.10"
 log = "0.4"
 once_cell = "1.17"
@@ -19,6 +18,6 @@ serde_yaml = "0.9"
 tokio = { version = "1.25", features = ["macros", "rt", "rt-multi-thread", "net", "signal", "sync"] }
 thiserror = "1.0"
 
-[[bin]]
+[lib]
 name = "guardity"
-path = "src/main.rs"
+path = "src/lib.rs"
